fix: improvements to scraping
- fixes #2690
