+++ package.json
@@ -58,7 +58,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@coralproject/bunyan-prettystream": "^0.1.4",
-    "@metascraper/helpers": "^5.7.14",
+    "@metascraper/helpers": "^5.7.21",
     "akismet-api": "^4.2.0",
     "apollo-server-express": "^2.8.1",
     "archiver": "^3.0.3",
@@ -106,11 +106,10 @@
     "lodash": "^4.17.15",
     "lru-cache": "^5.1.1",
     "luxon": "^1.12.0",
-    "metascraper-author": "^5.7.14",
-    "metascraper-date": "^5.7.14",
-    "metascraper-description": "^5.7.14",
-    "metascraper-image": "^5.7.14",
-    "metascraper-title": "^5.7.14",
+    "metascraper-author": "^5.8.0",
+    "metascraper-description": "^5.8.0",
+    "metascraper-image": "^5.8.0",
+    "metascraper-title": "^5.8.0",
     "mongodb": "^3.2.7",
     "mongodb-core": "^3.2.7",
     "ms": "^2.1.1",
