+++ src/core/server/services/stories/scraper/rules/modified.ts
@@ -1,12 +1,12 @@
-import { $jsonld } from "@metascraper/helpers";
+import { $jsonld, date, toRule } from "@metascraper/helpers";
 import { Rules } from "metascraper";
 
-import { wrap } from "./helpers";
+const toDate = toRule(date);
 
 export const modifiedScraper = (): Rules => ({
   modified: [
     // From: http://ogp.me/#type_article
-    wrap($jsonld("dateModified")),
-    wrap($ => $('meta[property="article:modified"]').attr("content")),
+    toDate($jsonld("dateModified")),
+    toDate($ => $('meta[property="article:modified"]').attr("content")),
   ],
 });
